+++ CHANGELOG.md
@@ -1,6 +1,6 @@
-# :fire: Improvements
+# :memo: Internal changes
 
-- Add default value to `types.reserve` [`58e0ed0`](https://github.com/keindev/mst-tools/commit/58e0ed09ea0540b2a9a672727a45025158572d28)
+- Fix `types.dispatcher` type [`ce63110`](https://github.com/keindev/mst-tools/commit/ce631103f1ea9cf7834207396bac100839914148)
 
 ---